Fixed issue with GraphGenerator related to deleting an edge that did not exist
Since I added the ability to parameterize the frequency of bidirectional graphs, there was a problem that occurred when trying delete an edge. This was caused by not having the code in the correct area. The problem was caused when an edge with a deadline of 0 was not added, we were still trying to delete it, which was causing a no key exception.