The notes for the current experiments had to be updated
The notes on the optimal solution for these instances had to be updated. When they were originally made, they were based on incorrect visualizations of the graphs. The cause of this was because the visualization function was not allowing both labels to coexist on the same edge (this happened when there were bidirectional edges).