Changed the timelimit for badinstance of edf test case
Initially we were deciding between t = 5 and t = 4 for the timelimit for the bad instance of edf. The choice to use 5 was based on the old opt algorithm that was not correctly producing the optimal solutions. I tested opt with 4 and it also served 4 requests, which is what we were thinking initially.