Added rough draft algorithm for EDF
A first approach for EDF is in this commit. There is a lot of debugging print statements in both OPT and EDF. This is mainly to show how each of the algorithms are behaving and what requests or jumps they are performing. It is notable that EDF may not be entirely accurate. Although it behaves as expected for the first test case, this does not mean that it is correct. Things to consider include instances where the window size is used or increased (release times need to be added for this). The instance used to test works because all requests are available since by default previously existing test cases have release times of 0. This means that the algorithm may not work properly when this is not the case.