Potential Issue with not updating after a jump
The available requests was not being updated if there was need for a jump. Because of this, deadlines that were equal to the current time were kept in available requests and could be served. This should not have been possible since the jump itself would be the currenttime being equal to the deadline, meaning we cannot serve it since the service time would be one after the deadline. This was fixed. However, I am noticing a new issue where the jump is not happening correctly. Specifically, there are instances where there is a jump made, but it is not necessary since the requests are continuous. This seems to be just a logic problem in the if statement, so it should be an easy fix in theory.