Seemingly working version of EDF
I think I have finally figured out an approach that correctly follows EDF. This version of the EDF implementation is currently limited to window sizes that do not need to change, this is because I currently do not have any existing test cases that do not have non-zero release times. I ran the existing test suite on EDF and the output seemingly matches with what is anticipated. There will likely be changes to this algorithm once the release times are implemented in the graph object. I also need to incorporate some way to alert when the solutions for both OPT and EDF are alarming (as stated in the notes). This seems like an if condition when I am running the tests, so I am not too worried about this for the moment. I mainly wanted to commit a version of EDF that I think is looking promising.