Changed time limit back to 5
I had recently changed the timelimit to 4, but it should be 5 for this test. Since the server starts at a non-request origin, this means that it takes the first time unit to get to the first request. I changed the actual code to represent this (< instead of <=) but the test case itself was meant to have a deadline of 5 not 4 (deadline of 4 means it will only serve 3 requests).