Potential fix to problem with serving requests exactly when the server gets there (mainly when deadline is 1)
I think that I did find an issue with being able to serve requests with a deadline of 1 since it took one time unit to get to the first request. The request itself would need to have a deadline of 2. I think changing the conditions to serve to be strcitly currentTime < time limit and currentTiem < deadline will fix this, along with other issues that could occur related to this.