Graph now can include release times
Release times have been included in the graph structure. I created a new function for adding edges. It is the main way to add edges that should be used when we want non-zero release times. The existing edge function for adding deadlines just assumes a zero release time (this was mainly to make sure existing graphs generators and test cases do not all need to be changed). I think the more challenging part will be the graph generation by text file input. I think that will be somewhat tricky to alter to support release times. I also need to now change opt and edf to respect (and recognize) release times.