generateGraphFromFile now supports adding release times
I have changed the way the function works. Now there are three possible ways edges can be added based on file input. If the number of components in the "edgeInfo" collection, then we have the first node, second node, release time, and deadline. If we only have 3 components, then we have first node, second node, and deadline. Otherwise we only have two components which are the first node and second node. This would represent a case where the edge should have an infinite deadline.