Fixed potential issue if input file does not have a time limit
If there was no given time limit in the last line, then the function would throw an exception. Because of this, I added a new try except statement to check if there is a last line with the time limit. If there isnt then the time limit will be given a default value of math.inf.