Graph visualization now works for edf and opt
The visualization function now has bigger labels, and it supports different coloring for edf and opt solutions. This was done by redrawing the actual requests served (parameter returned by the actual algorithm) in order to make this easier. EDF is yello and OPT is red. There is also a new helper function that is meant to force the timelimit to be the minimum value of the maximum deadline and the given timelimit. This is to ensure that no unnecessary iterations occur if we have small deadlines and are given a large timelimit. The timelimit itself was also added to the graph visualization as a title. This means that the timelimit is now visible with every graph that is visually printed.