EDF now supports recording of when the requests were served
The visualization function displays what time the served requests were served. EDF did not utilize this because I thought it would break the algorithm. It was actually very similar to how I did it with OPT. I basically created a collection that would be appended with the current time every time the current time was updated. Whenever a jump was made or no requests could be served but time still went on, then an 'x' was appended. These are then filtered out of the lists since this messes with the requests served collection whenever there are jumps (this is due to how I am getting each served request and its time).