Archived the old test cases
The old test cases have been used over and over. They also are no longer current in terms of not including deadlines of 0 or 1. They also all assume release times of 0, which is not as helpful for ensuring that the algorithms work when we have a nonzero release time. All archived test cases were moved into an archive folder. The new test cases will be made.