Made sure the windowSize is reset to 2 after each request is served.
If a request is not able to be served, then the windowSize is increased by one. The action of also increasing the currentTime by 1 each time this occurs effectively removes the need to increase the current time once a request is finally able to be served by how much the window size increased. Additionally, once a request is able to be served, the window size should be reset back to the starting initial value of 2.