EDF had a problem when the window size was updated
I added a boolean variable called inactive to be reset everytime the window had to be increased. This was mainly because when nothing could be served, the first servable request could be served at exactly the time it is released. I think this is meant to mimic that the server was just waiting at the node already to serve the request (no jump was needed during the idle time). Prior to this, it would move to the newly available request, then serve it. This means that if a request had a release time of 3, it would have a serve time of 4. Maybe this was correct, but it is definitely something to talk about. This version of EDF still does not use the windowsize when looking at which requests are available. It just increases the window size when nothing is available while also increasing the time. I think a better version would incorporate. I need to look more into the results of the test cases without release times of zero to see how the windowsize is working.