Changed graph popup so tests could run without having to close the popup
Everytime a graph is visualized, it is saved to the folder it exists in, but the popup of the graph visual then is showed and needs to be closed before the next case can start. This makes it impossible to have the tests running in the background. For that reason I simply commented out the part of the function that shows the graph.