Fixed OPT and EDF Timing
The timing for OPT and EDF was messed up because I was thinking of the time in terms of intervals. I first needed to change the current time to start at two, since this is the first time any possible request can be served (i.e. time 1 is spent traveling to any location so time 2 is when it can actually be served). The second thing that needed to be fixed was the bounds. The bound needed to be changed to <= since a request could be served at time 5 if its deadline was 5. Although this changed the solutions slightly in terms of the graph visualization, the actual optimal solution did not seem to change too much. Another big issue I found with edf was that the inclusion of deadlines meant the way the requests were sorted was wrong. Initially, all requests only had a deadline, so the lambda sorting function sorted it based on that one value for each key. Since there are now two values per key, the sorting was being done on the release time. This was the reason why the edf results were now incorrect. Once I realized this, I changed the sorting to be based on the second element, which is the deadline. This seemed to fix the main issue with EDF. I am more confident of EDF now, as I have tested it with old archived test cases, as well as some of the easier to manually computer new interesting test cases with nonzero release times.