Commit that fixes opt serving no solutions
When opt would have no release times that were close, it would report no solutions served. This was caused by the current request being updated regardless of if something was served or not. When the first available request had a large enough release time, this would result in zero requests served because the entire request list would be iterated through before anything could be served. This was fixed by only updating the current request if it was served. I am thinking it may be faster to just make the jump to the release time as long as it is within the timelimit. That will probably be in a future commit.