The function for running the test suite now saves and reports specific test cases to csv
I know one of the important things that needed to happen is that the test cases that were considered "interesting" needed to be somehow reported. I added the ability for these specific instances to be reported in a csv file. The purpose of this is to look back to the test cases and see the visuals of these reported instances.