Fixed problem with serving requests at timelimit
This commit fixes the issue related with requests being served at the time limit. We think that this is being caused because the condition for the while loop to keep iterating is currentTime<=timeLimit. I simply changed this for EDF and OPT to be strictly <. This seems to have fixed that problem. I also deleted the duplicate opt version I was working on that featured splitting the permutation into chunks. I think that the current way is promising and will work for larger permutations, therefore this second version for opt is not necessary (it also was not working). I made a helper function that writes to csv. This is where more flags can be created and it can be called to avoid duplicating this code for every test we run. I also changed the graph generator import statement to have an alias. This change was mainly because graph generator is long to type, and I am a fan of aliases.