Removed superfluous code and print statements
There were some variables left in graph that were not being used. The only one was colors, which was back from when I was trying to get the different coloring for opt and edf solutions. The way it is now does not utilize the variable so it is superfluous. There are also some print statements that were removed because they were from when I needed to test the helper function for getting the time limit. This works now, so the print statements are not needed.