Removed opt and edf related functions from here
I felt that the DarpExperimentation file was becomming too large. For that reason I have decided to make two new files for edf and opt related functions. Now DarpExperimentation file only contains code related to the experimentation process itself. This means this will be the main file that is run, and it will contain functions related to generating graphs that can be run along with csv reporting functions as well. I have already tested it to make sure it works still after making this big structural change.