New function to run randomly generated graphs
There is now a new helper function that can be used to randomly create graphs and save them to a folder after running OPT and EDF on them. This is very similar to the other function related to this. The main difference is that the other function uses graph generation method to convert text input into graphs. This function relies on the random graph generator instead.