Timing and fix to crashing problem when there are no solutions
I added a way for the run to be timed. The time is saved to csv along with the test name and solution. There was also a problem when no solutions were found. Initially, zero was returned when no solution is found, but the solution visualizer function requires values from a list of the solution. This is because when rides were served, additional data was returned as well. This was changed to make the code compatible for the case where there are no rides served.