Working version of opt
OPT was getting flagged sometimes for allegedly serving less than edf. After reviewing the instances where this happened, it became apparant that opt was the algorithm that was slightly off. The problem ended up being a couple of things. The first main problem I found was I was incorrectly iterating through the permutations. This caused only half of the permutations to be considered, resulting in occasional edf performing better than opt instances. The second issue found was that the current request was being incremented despite what occurred. This meant that if there was idle time, if the idle time was large enough then zero requests would be served since the whole list of requests would be iterated through. The third issue found was that edges with deadline 0 (non requests) were still in the list of edges. This resulted in not only more permutations, but orderings of list combinations that should not have been considered. In turn this resulted in incorrect solutions. After testing it with the new fixes, it seems opt is presumably working correctly again.