Fixed time representation and used permutations instead of islice to support larger request sizes
The time reflected was representing when the time the request was finished being served. For example, a request served between time [1,2] would have the time representation of 2. The code was changed to reflect the time representation to be when the service began. In terms of the changes made to the permutations, I found that islice reaches a max value since the range can no longer be represented as an int for large permutations. I then found out I could do without using islice, and use permutation to go through each permutation.