X character was still being appended to timing collection
I removed the code that filters out all of the 'x' characters that were initially used to represent a request not being served at that time (this was used for debugging purposes). I left one statement that still did this, so some solutions included that character, which should not be there.